rename classes, and references
